+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * `&Pageable` now has an `impl Pageable`
   * `api::paged` now supports `into_iter()` and `into_iter_async()` methods.
   * Add `api::runners::Runners` endpoint
+  * Add `api::runners::AllRunners` endpoint
 
 ## Breaking changes